refactored previous service_up? in ServicesController to Service model itself as self.up
this was to make an easier connection with othe methods that will be in the model along with
some added self methods (ie discovery). I have a basic working example of querying other servers
iteratively with curb to see what services are up. Would need to do this via threads
and make sure to only query other servers on their services app so they can query themselves
for their own list.