update recovery_state initializer to handle loading old cached state from database
Basically a saved state is still deamed fresh if its less than FRESH_FOR seconds
which is set inside the updated initializer. This file will check the current time
against the last saved database state's timestamp to see if it meets the FRESH_FOR
threshold. If it does, will cache the last saved state and output a message to the
server log indicating that choice. Otherwise it will simply output that the cache
has not been initialized. This will instead be handed by a cron job that syncs
the cache on a time interval, or during the first time a user hits a newly
spun up box (by calling Service.discovery).